Minor tweaks to the nvim tree /auto sessionfix.
Also, spent sometime messing with pylsp. It wasn't recognising my ROS2 plugins. Pretty sure the issue was it wasn
t sourcing my virtualenv? SO I moved the activate script over to ~/.profile so it is run even in non-interactive shells.

IDK, this feels confusing to me because I thought the ros2 libraries were stored in /opt/ros with the ros2 install...... in which case that wouldn't make a difference.

ok nah. I just removed that from my bash profile. And guess what.. rclpy definitions are still given. Looks like this issue is present in ubuntu at work and absent in arch at home :(