Fixed all my nvim-tree integration issues.
First we have to manually open the tree when we cwd to a folder with no existing session - that's becasue no_restore_cmds only works on VimEnter.

My other fix is that we have to close the tree every time we cwd because otherwise it will keep all of our old buffers alive when we switch to a new session.